requirements.txt
SQLAlchemy
PyYAML
mysql-connector-python
cx_Oracle
snowflake-connector-python
boto3
